+++ src/lib/pages/account-details/components/nfts/index.ts
@@ -1,2 +1,3 @@
 export * from "./NftsOverview";
-export * from "./NftsSection";
+export * from "./NftsSectionFull";
+export * from "./NftsSectionSequencer";